Create MyLibrary
This function will download the title and artist of songs in your library. There is a limit to how many songs you can pull in one run. its somewhere between 1000 and 2000. There are plenty of ways to get around this, but I'll try to post my way sometime in the near future.